app/requirements.txt
anyio==3.7.1
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.70.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pydantic==1.10.17
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==1.4.52
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.6
starlette==0.16.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.15.0
